requirements.txt
# DAC-3D System Dependencies

# Core
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Hardware Control
pypylon>=2.0.0  # Basler相机SDK
pyserial>=3.5   # 串口通信
pynq>=3.0.0     # FPGA控制(仅Zynq板卡需要)

# Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0  # 推理加速

# 3D Visualization
napari[all]>=0.4.18
vispy>=0.13.0
vtk>=9.2.0

# GUI
PyQt5>=5.15.9
pyqtgraph>=0.13.3

# Data Storage
h5py>=3.9.0
zarr>=2.16.0
tables>=3.8.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0
jsonschema>=4.18.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0
psutil>=5.9.0

# State Machine
transitions>=0.9.0

# Network
pyzmq>=25.1.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-qt>=4.2.0

# Code Quality
mypy>=1.4.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Utilities
tqdm>=4.65.0
colorama>=0.4.6
python-dotenv>=1.0.0
